refactor: move ws logic
